docs + bugfix on Edit button
I forgot to start a git repo when I started working on the project.. so its basically complete from my first commit. the only way I could remedy this would be to just rewrite it all.. which is tedious and really unnecessary.  I'll start committing from now on but I think I am almost finished the tasks for this midterm.